Consolidate time and tab stores into settings store
