+++ apps/antalmanac/src/components/Calendar/CourseCalendarEvent.tsx
@@ -15,7 +15,7 @@ import AppStore from '$stores/AppStore';
 import locationIds from '$lib/location_ids';
 import { useTabStore } from '$stores/TabStore';
 import { formatTimes } from '$stores/calendarizeHelpers';
-import { useTimeFormatStore } from '$stores/TimeStore';
+import { useTimeFormatStore } from '$stores/SettingsStore';
 
 const styles: Styles<Theme, object> = {
     courseContainer: {
